{arch,cpu}/x86, drivers/intel: Restore cpu_index error handling
Previously cpu_index() always succeeded, but since commit 095c931
(src/arch/x86: Use core apic id to get cpu_index()) it is now possible
for it to indicate an error by returning -1. This commit adds error
handling for all calls to cpu_index(), and restores several checks that
were removed in commit 7c712bb (Fix code that would trip -Wtype-limits)
but are now needed.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/32795
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/arch/x86/cpu.c
@@ -350,7 +350,7 @@ void arch_bootstate_coreboot_exit(void)
  * Hence new logic to use cpuid to fetch lapic id and matches with
  * cpus_default_apic_id[] variable to return correct cpu_index().
  */
-unsigned long cpu_index(void)
+int cpu_index(void)
 {
 	int i;
 	int lapic_id = initial_lapicid();
